+++ src/response/response.rs
@@ -2,7 +2,6 @@ use std::{
     cell::{Ref, RefMut},
     fmt,
     future::Future,
-    mem,
     pin::Pin,
     task::{Context, Poll},
 };
@@ -12,6 +11,7 @@ use actix_http::{
     http::{header::HeaderMap, StatusCode},
     Extensions, Response, ResponseHead,
 };
+use futures_core::ready;
 
 #[cfg(feature = "cookies")]
 use {
@@ -56,14 +56,6 @@ impl HttpResponse<Body> {
             error: Some(error),
         }
     }
-
-    // /// Convert response to response with body
-    // pub fn into_body<B>(self) -> HttpResponse<B> {
-    //     HttpResponse {
-    //         res: self.res.into_body(),
-    //         error: self.error,
-    //     }
-    // }
 }
 
 impl<B> HttpResponse<B> {
@@ -241,7 +233,7 @@ impl<B> HttpResponse<B> {
     // pub fn take_body(&mut self) -> ResponseBody<B> {
     //     self.res.take_body()
     // }
-    
+
     /// Extract response body
     pub fn into_body(self) -> B {
         self.res.into_body()
@@ -279,26 +271,28 @@ impl<B> From<HttpResponse<B>> for Response<B> {
 
         // TODO: expose cause somewhere?
         // if let Some(err) = res.error {
-        //     eprintln!("impl<B> From<HttpResponse<B>> for Response<B> let Some(err)");
-        //     return Response::from_error(err).into_body();
+        //     return Response::from_error(err);
         // }
 
         res.res
     }
 }
 
-impl Future for HttpResponse {
-    type Output = Result<Response<Body>, Error>;
+impl<B: Unpin> Future for HttpResponse<B> {
+    type Output = Result<Response<B>, Error>;
 
-    fn poll(mut self: Pin<&mut Self>, _: &mut Context<'_>) -> Poll<Self::Output> {
-        // if let Some(err) = self.error.take() {
-        //     return Poll::Ready(Ok(Response::from_error(err).into_body()));
-        // }
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+        if let Some(err) = self.error.take() {
+            return Poll::Ready(Err(err));
+        }
 
-        Poll::Ready(Ok(mem::replace(
-            &mut self.res,
-            Response::new(StatusCode::default()),
-        )))
+        let res = &mut self.res;
+        actix_rt::pin!(res);
+
+        match ready!(res.poll(cx)) {
+            Ok(val) => Poll::Ready(Ok(val)),
+            Err(err) => Poll::Ready(Err(err.into())),
+        }
     }
 }